Add font loading strategy for fast font viewing
My interpretation of the critical FOFT with data-uris. See:

https://www.zachleat.com/web/comprehensive-webfonts/

Some big improvements for time to screen are necessary: content is all in JS
right now, CSS is in a single render blocking file. But this sets our font
strategy, and will work great after those two larger issues are addressed.